+++ src/shared/ui/CodeEditor/languages/FuncLanguageDefinition.ts
