+++ src/shared/ui/CodeEditor/languages/index.ts
@@ -1,0 +1,2 @@
+export const TASM_LANGUAGE_ID = "tasm"
+export const FUNC_LANGUAGE_ID = "func"
